requirements.txt
requests
tqdm
Pillow
opencv-python
gallery-dl
beautifulsoup4
imutils
numba
psutil
imagehash
huggingface_hub
natsort
aiofiles
pytesseract